% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plots.R
\name{create_violin_plot}
\alias{create_violin_plot}
\title{Create Violin Plot with Custom Elements, P-Values, and Flexible Options}
\usage{
create_violin_plot(
  data,
  group_column,
  value_column,
  x_labels = NULL,
  group_colors,
  breaks = NULL,
  limits = NULL,
  outliers_color = "red",
  x_lab = "",
  y_lab = NULL,
  p_value = TRUE,
  p_value_format = "asterisk",
  box_plot = TRUE,
  outliers = TRUE,
  median = TRUE
)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{group_column}{A string representing the column name that defines the groups (x-axis).}

\item{value_column}{A string representing the column name that contains the values (y-axis).}

\item{x_labels}{A character vector representing the custom labels for the x-axis groups. Default is `NULL`.}

\item{group_colors}{A named character vector of colors, where the names correspond to the group levels and values correspond to the colors.}

\item{breaks}{An optional numeric vector specifying the breaks on the y-axis. Default is `NULL`.}

\item{limits}{An optional numeric vector of length two specifying the limits on the y-axis. Default is `NULL`.}

\item{outliers_color}{A string representing the color of the outliers. Default is "red".}

\item{x_lab}{A string representing the label for the x-axis (default: "").}

\item{y_lab}{A string representing the label for the y-axis (default: the `value_column` name).}

\item{p_value}{Logical, whether to include p-values in the plot. Default is TRUE.}

\item{p_value_format}{A string, either "asterisk" or "numeric", to control how p-values are shown. Default is "asterisk".}

\item{box_plot}{Logical, whether to include a boxplot inside the violin plot. Default is TRUE.}

\item{outliers}{Logical, whether to include outliers on the plot. Default is TRUE.}

\item{median}{Logical, whether to include the median point on the plot. Default is TRUE.}
}
\value{
A `ggplot2` object representing the violin plot.
}
\description{
This function generates a violin plot with additional elements like boxplots, medians, outliers,
and p-values for pairwise group comparisons. Users can optionally set y-axis breaks, limits, outlier color,
axis labels, and control the inclusion of p-values, boxplots, and medians.
}
\examples{
# Example usage with mock data
data <- data.frame(
  group = rep(c("A", "B", "C"), each = 100),
  value = c(rnorm(100, mean = 0), rnorm(100, mean = 1), rnorm(100, mean = 2))
)
group_colors <- c('A' = "#8DA0CB80", 'B' = "#66C2A580", 'C' = "#E41A1C80")
p <- create_violin_plot(data, "group", "value", group_colors = group_colors, y_lab = bquote(bold(Delta ~ Log(activity))))
print(p)
}
