% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_violin_plot.R
\name{split_violin_plot}
\alias{split_violin_plot}
\title{Create a Split Violin Plot with Custom Elements and Options}
\usage{
split_violin_plot(
  data,
  group_column,
  value_column,
  split_column,
  colors,
  labels = NULL,
  x_lab = "Groups",
  y_lab = value_column,
  breaks = NULL,
  limits = NULL,
  outliers = TRUE,
  CI = TRUE,
  median = TRUE,
  n_obs = TRUE
)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{group_column}{A string representing the name of the column that defines the groups (x-axis).}

\item{value_column}{A string representing the name of the column containing the values (y-axis).}

\item{split_column}{A string representing the name of the column used to split the violins.}

\item{colors}{A named vector of two colors to be used for the left and right halves of the violins.}

\item{labels}{An optional named vector of labels for the legend corresponding to the `split` values.}

\item{x_lab}{A string representing the label for the x-axis (default: "Groups").}

\item{y_lab}{A string representing the label for the y-axis (default: the `value_column` name).}

\item{breaks}{An optional vector of breaks for the y-axis.}

\item{limits}{An optional vector of length two defining the limits of the y-axis.}

\item{outliers}{Logical, whether to include outliers (default TRUE).}

\item{CI}{Logical, whether to include confidence intervals (default TRUE).}

\item{median}{Logical, whether to include medians (default TRUE).}

\item{n_obs}{Logical, whether to include number of observations (default TRUE).}
}
\value{
A ggplot2 object with the split violin plot and custom legend.
}
\description{
This function generates split violin plots with custom elements such as Q1, Q3, medians,
confidence intervals, outliers, and observation counts. It allows for flexible inclusion or exclusion
of these elements, and customization of axis labels.
}
\examples{
# Example data
data <- data.frame(
  group = rep(c("A", "B", "C"), each = 100),
  value = c(rnorm(100, mean = 0), rnorm(100, mean = 1), rnorm(100, mean = 2)),
  split_criteria = sample(c(TRUE, FALSE), 300, replace = TRUE)
)

colors <- c("TRUE" = "#FF9999", "FALSE" = "#99CCFF")
labels <- c("TRUE" = "first group", "FALSE" = "second group")

# Generate the split violin plot with custom labels
p <- split_violin_plot(data, "group", "value", "split_criteria", colors, labels = labels,
                       x_lab = "Groups", y_lab = bquote(bold(Delta ~ Log(activity))), outliers = TRUE, CI = TRUE, median = TRUE, n_obs = TRUE)
print(p)
}
