% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plots_with_comparison_group.R
\name{violin_plots_with_comparison_group}
\alias{violin_plots_with_comparison_group}
\title{Create Violin Plots with a Comparison Group}
\usage{
violin_plots_with_comparison_group(
  data,
  group_column,
  value_column,
  comparison_group,
  basic_color = "#8DA0CB80",
  comparison_color = "#66C2A580",
  x_labels = NULL,
  breaks = NULL,
  limits = NULL,
  outliers_color = "red",
  x_lab = "",
  y_lab = NULL,
  p_value = TRUE,
  p_value_format = "asterisk",
  box_plot = TRUE,
  outliers = TRUE,
  median = TRUE
)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{group_column}{A string representing the name of the column that defines the groups (x-axis).}

\item{value_column}{A string representing the name of the column containing the values (y-axis).}

\item{comparison_group}{A string representing the name of the group to be used as the comparison group.}

\item{basic_color}{A string representing the color for the non-comparison groups. Default is "#8DA0CB80".}

\item{comparison_color}{A string representing the color for the comparison group. Default is "#66C2A580".}

\item{x_labels}{A character vector representing the custom labels for the x-axis groups. Default is `NULL`.}

\item{breaks}{An optional numeric vector specifying the breaks on the y-axis. Default is `NULL`.}

\item{limits}{An optional numeric vector of length two specifying the limits on the y-axis. Default is `NULL`.}

\item{outliers_color}{A string representing the color of the outliers. Default is "red".}

\item{x_lab}{A string representing the label for the x-axis (default: "").}

\item{y_lab}{A string representing the label for the y-axis (default: the `value_column` name).}

\item{p_value}{Logical, whether to include p-values in the plot. Default is TRUE.}

\item{p_value_format}{A string, either "asterisk" or "numeric", to control how p-values are shown. Default is "asterisk".}

\item{box_plot}{Logical, whether to include a boxplot inside the violin plot. Default is TRUE.}

\item{outliers}{Logical, whether to include outliers on the plot. Default is TRUE.}

\item{median}{Logical, whether to include the median point on the plot. Default is TRUE.}
}
\value{
A `ggplot2` object representing the violin plot.
}
\description{
This function creates violin plots for different groups, with a specific comparison group
highlighted. The comparison group is displayed in a different color, and the function can
optionally calculate p-values for comparisons between the other groups and the comparison group.
}
\examples{
# Example usage with mock data
data <- data.frame(
  group = rep(c("A", "B", "C"), each = 100),
  value = c(rnorm(100, mean = 0), rnorm(100, mean = 1), rnorm(100, mean = 2))
)
p <- violin_plots_with_comparison_group(data, "group", "value", comparison_group = "A")
print(p)
}
